Deferred call / async msg context error (#4869)
