+++ massa-api/src/tests/public.rs
@@ -634,6 +634,8 @@ async fn get_filtered_sc_output_event() {
                     ),
                     is_final: false,
                     is_error: false,
+                    deferred_call_id: None,
+                    async_msg_id: None,
                 },
                 data: "massa".to_string(),
             }]
