+++ massa-execution-worker/src/interface_impl.rs
@@ -7,10 +7,10 @@
 
 use crate::context::ExecutionContext;
 use anyhow::{anyhow, bail, Result};
-use massa_async_pool::{AsyncMessage, AsyncMessageTrigger};
 use massa_deferred_calls::DeferredCall;
 use massa_execution_exports::ExecutionConfig;
 use massa_execution_exports::ExecutionStackElement;
+use massa_models::async_msg::{AsyncMessage, AsyncMessageTrigger};
 use massa_models::bytecode::Bytecode;
 use massa_models::datastore::get_prefix_bounds;
 use massa_models::deferred_calls::DeferredCallId;
